Move helper functions to utils
Refs PV-428
