+++ parking_permits/utils.py
@@ -133,3 +133,31 @@ def get_permit_prices(
             }
         )
     return permit_prices
+
+
+def get_meta_value(meta_pair_list, meta_pair_key):
+    return next(
+        (
+            meta_pair.get("value")
+            for meta_pair in meta_pair_list
+            if meta_pair.get("key") == meta_pair_key
+        ),
+        None,
+    )
+
+
+def snake_to_camel_dict(dictionary):
+    res = dict()
+    for key in dictionary.keys():
+        if isinstance(dictionary[key], dict):
+            res[camel_str(key)] = snake_to_camel_dict(dictionary[key])
+        elif isinstance(dictionary[key], list):
+            res[camel_str(key)] = [snake_to_camel_dict(val) for val in dictionary[key]]
+        else:
+            res[camel_str(key)] = dictionary[key]
+    return res
+
+
+def camel_str(snake_str):
+    first, *others = snake_str.split("_")
+    return "".join([first.lower(), *map(str.title, others)])
